requirements.txt
#install autogen using the terminal
# pip install git+https://github.com/microsoft/autogen.git 
pyautogen[autobuild]